# PaperLess.Frontend/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Mock Server",
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
...
    "/api/documents" : {
      "get" : {
        "operationId" : "GetDocuments",
        "parameters" : [ {
          "explode" : true,
          "in" : "query",
          "name" : "Page",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "form"
        }, {
          "explode" : true,
          "in" : "query",
          "name" : "page_size",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "form"
        }, {
          "explode" : true,
          "in" : "query",
          "name" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          },
          "style" : "form"
        }, {
          "explode" : true,
          "in" : "query",
          "name" : "ordering",
          "required" : false,
          "schema" : {
            "type" : "string"
          },
          "style" : "form"
        }, {
          "explode" : true,
          "in" : "query",
          "name" : "tags__id__all",
          "required" : false,
          "schema" : {
            "items" : {
              "format" : "int32",
              "type" : "integer"
            },
            "type" : "array"
          },
          "style" : "form"
        }, {
          "explode" : true,
          "in" : "query",
          "name" : "document_type__id",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "form"
        }, {
          "explode" : true,
          "in" : "query",
          "name" : "correspondent__id",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "form"
        }, {
          "explode" : true,
          "in" : "query",
          "name" : "truncate_content",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          },
          "style" : "form"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        },
        "tags" : [ "Documents" ]
      }
    },
    "/api/documents/{id}" : {
      "delete" : {
        "operationId" : "DeleteDocument",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        },
        "tags" : [ "Documents" ]
      },
      "put" : {
        "operationId" : "UpdateDocument",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "simple"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Document"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Document"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Document"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        },
        "tags" : [ "Documents" ]
      }
    },
    "/api/documents/post_document" : {
      "post" : {
        "operationId" : "UploadDocument",
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "encoding" : {
                "title" : {
                  "style" : "form"
                },
                "created" : {
                  "style" : "form"
                },
                "document_type" : {
                  "style" : "form"
                },
                "tags" : {
                  "style" : "form"
                },
                "correspondent" : {
                  "style" : "form"
                },
                "document" : {
                  "style" : "form"
                }
              },
              "schema" : {
                "$ref" : "#/components/schemas/UploadDocument_request"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        },
        "tags" : [ "Documents" ]
      }
    },
    "/api/documents/{id}/thumb" : {
      "get" : {
        "operationId" : "GetDocumentThumb",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        },
        "tags" : [ "Documents" ]
      }
    },
    "/api/documents/{id}/download" : {
      "get" : {
        "operationId" : "DownloadDocument",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "simple"
        }, {
          "explode" : true,
          "in" : "query",
          "name" : "original",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          },
          "style" : "form"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        },
        "tags" : [ "Documents" ]
      }
    },
    "/api/documents/{id}/preview" : {
      "get" : {
        "operationId" : "GetDocumentPreview",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        },
        "tags" : [ "Documents" ]
      }
    },
    "/api/documents/{id}/metadata" : {
      "get" : {
        "operationId" : "GetDocumentMetadata",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        },
        "tags" : [ "Documents" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Document" : {
        "additionalProperties" : false,
        "example" : {
          "archive_serial_number" : "archive_serial_number",
          "added" : "2000-01-23T04:56:07.000+00:00",
          "created" : "2000-01-23T04:56:07.000+00:00",
          "title" : "title",
          "content" : "content",
          "tags" : [ 5, 5 ],
          "storage_path" : 5,
          "archived_file_name" : "archived_file_name",
          "modified" : "2000-01-23T04:56:07.000+00:00",
          "correspondent" : 6,
          "original_file_name" : "original_file_name",
          "id" : 0,
          "created_date" : "2000-01-23T04:56:07.000+00:00",
          "document_type" : 1
        },
        "properties" : {
          "id" : {
            "format" : "int32",
            "type" : "integer"
          },
          "correspondent" : {
            "format" : "int32",
            "nullable" : true,
            "type" : "integer"
          },
          "document_type" : {
            "format" : "int32",
            "nullable" : true,
            "type" : "integer"
          },
          "storage_path" : {
            "format" : "int32",
            "nullable" : true,
            "type" : "integer"
          },
          "title" : {
            "nullable" : true,
            "type" : "string"
          },
          "content" : {
            "nullable" : true,
            "type" : "string"
          },
          "tags" : {
            "items" : {
              "format" : "int32",
              "type" : "integer"
            },
            "nullable" : true,
            "type" : "array"
          },
          "created" : {
            "format" : "date-time",
            "type" : "string"
          },
          "created_date" : {
            "format" : "date-time",
            "type" : "string"
          },
          "modified" : {
            "format" : "date-time",
            "type" : "string"
          },
          "added" : {
            "format" : "date-time",
            "type" : "string"
          },
          "archive_serial_number" : {
            "nullable" : true,
            "type" : "string"
          },
          "original_file_name" : {
            "nullable" : true,
            "type" : "string"
          },
          "archived_file_name" : {
            "nullable" : true,
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "UploadDocument_request" : {
        "properties" : {
          "title" : {
            "type" : "string"
          },
          "created" : {
            "format" : "date-time",
            "type" : "string"
          },
          "document_type" : {
            "format" : "int32",
            "type" : "integer"
          },
          "tags" : {
            "items" : {
              "format" : "int32",
              "type" : "integer"
            },
            "type" : "array"
          },
          "correspondent" : {
            "format" : "int32",
            "type" : "integer"
          },
          "document" : {
            "items" : {
              "format" : "binary",
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      }
    }
  }
}